requirements.txt
beautifulsoup4==4.11.2
matplotlib==3.7.1
numpy==1.24.1
Pillow==9.4.0
requests==2.28.2
scipy==1.10.1
wordcloud==1.8.2.2
